Update 3 x n 타일링.py

+++ 프로그래머스/lv2/12902. 3 x n 타일링/3 x n 타일링.py
@@ -1,6 +1,6 @@
 def solution(n):
     res = [0, 3, 11]
-    idx = int(n/2)
+    idx = n/2
     
     if n % 2 != 0: return 0
